+++ src/main/java/org/jevis/jeconfig/batchmode/SensorMap.java
@@ -26,95 +26,81 @@ import org.apache.commons.io.IOUtils;
  */
 public class SensorMap {
 
-    
     static InputStream inputStream;
+
     public SensorMap(String urlToRead) {
 
-     URL url;
-     HttpURLConnection conn;
-     Logger.getLogger(SensorMap.class.getName()).log(Level.INFO, "Host name: " + urlToRead);
+        URL url;
+        HttpURLConnection conn;
+        Logger.getLogger(SensorMap.class.getName()).log(Level.INFO, "Host name: " + urlToRead);
 
-     try {    
-         url = new URL(urlToRead);
-         conn = (HttpURLConnection) url.openConnection();
-         conn.setRequestMethod("GET");
-         this.inputStream = conn.getInputStream(); 
+        try {
+            url = new URL(urlToRead);
+            conn = (HttpURLConnection) url.openConnection();
+            conn.setRequestMethod("GET");
+            this.inputStream = conn.getInputStream();
 
-     } catch (MalformedURLException ex) {       
-         Logger.getLogger(SensorMap.class.getName()).log(Level.SEVERE,null, ex);
-     } catch (ProtocolException ex) {
-         Logger.getLogger(SensorMap.class.getName()).log(Level.SEVERE,null, ex);
-     } catch (IOException ex) {
-         Logger.getLogger(SensorMap.class.getName()).log(Level.SEVERE,null, ex);
-     }   
-    }
-    
-   
-    
-    
-  
- public static void main(String[] args) {
-
-         SensorMap sensorMapClass = new SensorMap("http://localhost:3001/db");
-         Map<String, LinkedHashMap<String, String>> sm = sensorMapClass.getSensorMap(inputStream);
-         
-         
-         for (Map.Entry<String, LinkedHashMap<String, String>> sensorEntry : sm.entrySet()) {
-            if(sensorEntry.getValue().size()!=3){
-                continue;
-            }
-            String sensorName = sensorEntry.getKey();
-            String unit = (String) sensorEntry.getValue().keySet().toArray()[1];
-            System.out.println(sensorName + " + " + unit);
-         }
-    }  
- 
-    /**
-     * 
-     * @param inputStream json string
-     * @return Map<String, Map<String, String>> returns a map with sensor meta data
-     * 
-     * e.g. 
-     * "T": {
-     *      "MacAddr": "28524295050000AB",
-     *      "Temperature": "sd.22",
-     *      "Time": "2015-08-09T14:27:43+02:00"
-     *  },
-     *  "RH": {
-     *      "MacAddr": "28524295050000AB",
-     *      "Relative Humidity": "33.33",
-     *      "Time": "2015-08-09T14:27:43+02:00"
-     *  }
-     * 
-     *  for (String key : map.keySet()) {
-            
+        } catch (MalformedURLException ex) {
+            Logger.getLogger(SensorMap.class.getName()).log(Level.SEVERE, null, ex);
+        } catch (ProtocolException ex) {
+            Logger.getLogger(SensorMap.class.getName()).log(Level.SEVERE, null, ex);
+        } catch (IOException ex) {
+            Logger.getLogger(SensorMap.class.getName()).log(Level.SEVERE, null, ex);
         }
-     * use the following code to get sensor name and unit
-     * 
-     * for (Map.Entry<String, LinkedHashMap<String, String>> sensorEntry : map.entrySet()) {
-            if(sensorEntry.getValue().size()!=3){
+    }
+
+    public static void main(String[] args) {
+
+        SensorMap sensorMapClass = new SensorMap("http://localhost:3001/db");
+        Map<String, LinkedHashMap<String, String>> sm = sensorMapClass.getSensorMap(inputStream);
+
+        for (Map.Entry<String, LinkedHashMap<String, String>> sensorEntry : sm.entrySet()) {
+            if (sensorEntry.getValue().size() != 3) {
                 continue;
             }
             String sensorName = sensorEntry.getKey();
             String unit = (String) sensorEntry.getValue().keySet().toArray()[1];
             System.out.println(sensorName + " + " + unit);
-         }
-     * 
-     */
-    public Map<String, LinkedHashMap<String, String>> getSensorMap(InputStream inputStream){
-  
-     Map<String,LinkedHashMap<String, String>> map = null;
-     try {
-         String jsonString = IOUtils.toString(inputStream);
-         //String jsonTestString ="{\"T\": {  \"MacAddr\": \"28524295050000AB\",  \"Temperature\": \"24.19\",  \"Time\": \"2014-05-26T07:06:43+02:00\"}}";
-         Gson gson = new Gson();
-         Type complexMap = new TypeToken<Map<String,LinkedHashMap<String, String>>>(){}.getType();
-         map = gson.fromJson(jsonString, complexMap); 
-     } catch (IOException ex) {
-         Logger.getLogger(SensorMap.class.getName()).log(Level.SEVERE, null, ex);
-     }
-
-         return map;
+        }
     }
-    
+
+    /**
+     *
+     * @param inputStream json string
+     * @return Map<String, Map<String, String>> returns a map with sensor meta
+     * data
+     *
+     * e.g. "T": { "MacAddr": "28524295050000AB", "Temperature": "sd.22",
+     * "Time": "2015-08-09T14:27:43+02:00" }, "RH": { "MacAddr":
+     * "28524295050000AB", "Relative Humidity": "33.33", "Time":
+     * "2015-08-09T14:27:43+02:00" }
+     *
+     * for (String key : map.keySet()) {
+     *
+     * }
+     * use the following code to get sensor name and unit
+     *
+     * for (Map.Entry<String, LinkedHashMap<String, String>> sensorEntry :
+     * map.entrySet()) { if(sensorEntry.getValue().size()!=3){ continue; }
+     * String sensorName = sensorEntry.getKey(); String unit = (String)
+     * sensorEntry.getValue().keySet().toArray()[1];
+     * System.out.println(sensorName + " + " + unit); }
+     *
+     */
+    public Map<String, LinkedHashMap<String, String>> getSensorMap(InputStream inputStream) {
+
+        Map<String, LinkedHashMap<String, String>> map = null;
+        try {
+            String jsonString = IOUtils.toString(inputStream);
+            //String jsonTestString ="{\"T\": {  \"MacAddr\": \"28524295050000AB\",  \"Temperature\": \"24.19\",  \"Time\": \"2014-05-26T07:06:43+02:00\"}}";
+            Gson gson = new Gson();
+            Type complexMap = new TypeToken<Map<String, LinkedHashMap<String, String>>>() {
+            }.getType();
+            map = gson.fromJson(jsonString, complexMap);
+        } catch (IOException ex) {
+            Logger.getLogger(SensorMap.class.getName()).log(Level.SEVERE, null, ex);
+        }
+
+        return map;
+    }
 }
